add resource and API

+++ app/Http/Controllers/WorkerController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\Worker\IndexRequest;
 use App\Http\Requests\Worker\StoreRequest;
 use App\Http\Requests\Worker\UpdateRequest;
 use App\Models\Worker;
+use Illuminate\Http\Request;
 
 class WorkerController extends Controller
 {
@@ -35,7 +36,7 @@ class WorkerController extends Controller
             $workersQuery->where('is_married', $data['is_married']);
         }
 
-        $workers = $workersQuery->paginate(2);
+        $workers = $workersQuery->paginate(10);
 
 
         return view('worker.index', compact('workers'));
@@ -47,44 +48,63 @@ class WorkerController extends Controller
         return view('worker.show', compact('worker'));
     }
 
-    public function create()
+    public function create(Request $request)
     {
+        if ($request->user()->cannot('create', Worker::class)) {
+            abort(403);
+        }
         return view('worker.create');
     }
 
     public function store(StoreRequest $request)
     {
 
+        if ($request->user()->cannot('create', Worker::class)) {
+            abort(403);
+        }
+
         $validated = $request->validated();
         $worker = Worker::create($validated);
 
-        return redirect()->route('worker.index')->with('success', 'Worker has been created');
+        return redirect()->route('workers.index')->with('success', 'Worker has been created');
 
     }
 
-    public function edit(Worker $worker) {
+    public function edit(Worker $worker , Request $request) {
+
+        if ($request->user()->cannot('update', $worker)) {
+            abort(403);
+        }
 
         return view('worker.edit', compact('worker'));
 
     }
 
-    public function update(UpdateRequest $request, $id) {
+    public function update(UpdateRequest $request, Worker $worker) {
 
-        $worker = Worker::findOrFail($id);
+        if ($request->user()->cannot('update', $worker)) {
+            abort(403);
+        }
+
+        $worker = Worker::findOrFail($worker->id);
 
         $worker->update($request->validated());
 
-        return redirect()->route('worker.index')->with('success', 'Worker has been updated');
+        return redirect()->route('workers.index')->with('success', 'Worker has been updated');
 
     }
 
-    public function delete($id) {
+    public function destroy(Worker $worker , Request $request) {
 
-        $worker = Worker::find($id);
+        if ($request->user()->cannot('delete', $worker )) {
+            abort(403);
+        }
+
+        $worker = Worker::find($worker->id);
 
         $worker->delete();
 
-        return redirect()->route('worker.index')->with('success', 'Worker has been deleted');
+        return redirect()->route('workers.index')->with('success', 'Worker has been deleted');
 
     }
 }
